+++ FriendApi.lua
@@ -54,7 +54,7 @@ end
 __Static__() __AutoCache__()
 function AshBlzSkinApi.UnitNameColor()
     return Wow.FromUnitEvent(Wow.FromEvent("UNIT_NAME_UPDATE", "GROUP_ROSTER_UPDATE")):Next():Map(function(unit)
-        if not DB.Appearance.Name.FriendsNameColoring or  not UnitIsPlayer(unit) or UnitIsUnit("player", unit) then return Color.WHITE end
+        if not DB().Appearance.Name.FriendsNameColoring or  not UnitIsPlayer(unit) or UnitIsUnit("player", unit) then return Color.WHITE end
         local guid = UnitGUID(unit)
         if tContains(BNFriendList, guid) then
             return Color.BATTLENET